// types.ts
export interface HeartData {
  id: number;
  left: string;
  size: string;
  duration: string;
  delay: string;
  color: string;
}

export interface SparkleData {
  id: number;
  left: string;
  top: string;
  duration: string;
}

export interface PhotoMemory {
  id: number;
  url: string;
  caption: string;
}